requirements.txt
annotated-types==0.6.0
anyio==3.7.1
beautifulsoup4==4.12.2
bidict==0.22.1
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
distro==1.8.0
Flask==3.0.0
Flask-SocketIO==5.3.6
gevent==23.9.1
gevent-websocket==0.10.1
google-search-results==2.4.2
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Markdown==3.5.1
MarkupSafe==2.1.3
openai==1.3.5
psycopg2==2.9.9
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
pymdown-extensions==10.3.1
python-dateutil==2.8.2
python-engineio==4.8.0
python-geoip-python3==1.3
python-socketio==5.10.0
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
tiktoken==0.5.1
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.7
uWSGI==2.0.23
Werkzeug==3.0.1
wsproto==1.2.0
zope.event==5.0
zope.interface==6.1
